# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example",
        "description": "Example API Documentation",
        "contact": {
            "name": "Example",
            "email": "example@test.com"
        },
        "version": "0.1"
    },
    "paths": {
...
        "/api/regions/gallery/postList": {
            "get": {
                "tags": [
                    "Gallery Posts"
                ],
                "summary": "Get paginated list of gallery posts",
                "description": "Retrieve a paginated list of posts for a specific gallery based on gallery_id.",
                "operationId": "getGalleryPosts",
                "parameters": [
                    {
                        "name": "gallery_id",
                        "in": "query",
                        "description": "ID of the gallery for which posts are to be retrieved",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response containing paginated gallery posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "description": "Current page number",
                                            "type": "integer"
                                        },
                                        "data": {
                                            "description": "List of gallery posts",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Post ID",
                                                        "type": "integer"
                                                    },
                                                    "gallery_id": {
                                                        "description": "Gallery ID",
                                                        "type": "integer"
                                                    },
                                                    "user_id": {
                                                        "description": "User ID of the post author",
                                                        "type": "integer"
                                                    },
                                                    "user_name": {
                                                        "description": "Username of the post author",
                                                        "type": "string"
                                                    },
                                                    "title": {
                                                        "description": "Title of the post",
                                                        "type": "string"
                                                    },
                                                    "content": {
                                                        "description": "Content of the post",
                                                        "type": "string"
                                                    },
                                                    "views": {
                                                        "description": "View count of the post",
                                                        "type": "integer"
                                                    },
                                                    "created_at": {
                                                        "description": "Post creation timestamp",
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "description": "Post update timestamp",
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "description": "URL of the first page",
                                            "type": "string"
                                        },
                                        "last_page": {
                                            "description": "Last page number",
                                            "type": "integer"
                                        },
                                        "last_page_url": {
                                            "description": "URL of the last page",
                                            "type": "string"
                                        },
                                        "next_page_url": {
                                            "description": "URL of the next page",
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "prev_page_url": {
                                            "description": "URL of the previous page",
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "total": {
                                            "description": "Total number of posts",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "부적절한 접근입니다."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "No posts found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "게시글이 존재하지 않습니다"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}